+++ src/parser/type/heterogenous/short_string_mixed_type.cpp
@@ -17,7 +17,7 @@
 namespace infinity {
 
 bool ShortStrMixedType::Equal(const std::string &input) {
-    if (length != input.length())
+    if (length != (int8_t)input.length())
         return false;
     for (size_t i = 0; const auto &c : input) {
         if (this->ptr[i++] != c)
